refactor: split group screen into components
