+++ app/(tabs)/groups.tsx
@@ -1,268 +1,23 @@
 import { Button } from "@/components/Button";
 import { EmptyState } from "@/components/EmptyState";
+import { AddFriendModal } from "@/components/groups/AddFriendModal";
+import { CreateGroupModal } from "@/components/groups/CreateGroupModal";
+import { FriendItem } from "@/components/groups/FriendItem";
+import { GroupItem } from "@/components/groups/GroupItem";
+import { SearchBar } from "@/components/groups/SearchBar";
 import { colors } from "@/constants/Colors";
 import { Friend, useFriendsStore } from "@/hooks/useFriendsStore";
 import { Group, useGroupsStore } from "@/hooks/useGroupsStore";
 import { useRouter } from "expo-router";
-import {
-  ChevronRight,
-  Plus,
-  Receipt,
-  Search,
-  User,
-  Users,
-  X,
-} from "lucide-react-native";
+import { Plus, User, Users } from "lucide-react-native";
 import React, { useState } from "react";
 import {
-  Alert,
   FlatList,
-  Keyboard,
-  KeyboardAvoidingView,
-  Modal,
-  Platform,
   StyleSheet,
   Text,
-  TextInput,
   TouchableOpacity,
-  TouchableWithoutFeedback,
   View,
 } from "react-native";
-
-type CreateGroupModalProps = {
-  visible: boolean;
-  onClose: () => void;
-  onCreateGroup: (name: string) => void;
-};
-
-const CreateGroupModal = ({
-  visible,
-  onClose,
-  onCreateGroup,
-}: CreateGroupModalProps) => {
-  const [newGroupName, setNewGroupName] = useState("");
-
-  const handleCreate = () => {
-    if (!newGroupName.trim()) {
-      Alert.alert("Error", "Please enter a group name");
-      return;
-    }
-
-    onCreateGroup(newGroupName.trim());
-    setNewGroupName("");
-  };
-
-  return (
-    <Modal
-      visible={visible}
-      transparent
-      animationType="slide"
-      onRequestClose={onClose}
-    >
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : "height"}
-        style={styles.modalContainer}
-      >
-        <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
-          <View style={styles.modalContent}>
-            <View style={styles.modalHeader}>
-              <Text style={styles.modalTitle}>Create New Group</Text>
-              <TouchableOpacity style={styles.closeButton} onPress={onClose}>
-                <X size={20} color={colors.textSecondary} />
-              </TouchableOpacity>
-            </View>
-
-            <View style={styles.inputContainer}>
-              <Text style={styles.inputLabel}>Group Name</Text>
-              <TextInput
-                style={styles.input}
-                value={newGroupName}
-                onChangeText={setNewGroupName}
-                placeholder="Enter group name"
-                placeholderTextColor={colors.textLight}
-                autoFocus
-              />
-            </View>
-
-            <View style={styles.modalActions}>
-              <Button
-                title="Cancel"
-                onPress={onClose}
-                variant="outline"
-                containerStyle={{ flex: 1, marginHorizontal: 4 }}
-              />
-              <Button
-                title="Create"
-                onPress={handleCreate}
-                containerStyle={{ flex: 1, marginHorizontal: 4 }}
-              />
-            </View>
-          </View>
-        </TouchableWithoutFeedback>
-      </KeyboardAvoidingView>
-    </Modal>
-  );
-};
-
-type AddFriendModalProps = {
-  visible: boolean;
-  onClose: () => void;
-  onAddFriend: (name: string, phone: string) => void;
-};
-
-const AddFriendModal = ({
-  visible,
-  onClose,
-  onAddFriend,
-}: AddFriendModalProps) => {
-  const [newFriendName, setNewFriendName] = useState("");
-  const [newFriendPhone, setNewFriendPhone] = useState("");
-
-  const handleAdd = () => {
-    if (!newFriendName.trim()) {
-      Alert.alert("Error", "Please enter a friend name");
-      return;
-    }
-
-    if (!newFriendPhone.trim()) {
-      Alert.alert("Error", "Please enter a phone number");
-      return;
-    }
-
-    onAddFriend(newFriendName.trim(), newFriendPhone.trim());
-    setNewFriendName("");
-    setNewFriendPhone("");
-  };
-
-  return (
-    <Modal
-      visible={visible}
-      transparent
-      animationType="slide"
-      onRequestClose={onClose}
-    >
-      <View style={styles.modalContainer}>
-        <View style={styles.modalContent}>
-          <View style={styles.modalHeader}>
-            <Text style={styles.modalTitle}>Add New Friend</Text>
-            <TouchableOpacity style={styles.closeButton} onPress={onClose}>
-              <X size={20} color={colors.textSecondary} />
-            </TouchableOpacity>
-          </View>
-
-          <View style={styles.inputContainer}>
-            <Text style={styles.inputLabel}>Friend&apos;s Name</Text>
-            <TextInput
-              style={styles.input}
-              value={newFriendName}
-              onChangeText={setNewFriendName}
-              placeholder="Enter friend's name"
-              placeholderTextColor={colors.textLight}
-              autoFocus
-            />
-          </View>
-
-          <View style={styles.inputContainer}>
-            <Text style={styles.inputLabel}>Phone Number</Text>
-            <TextInput
-              style={styles.input}
-              value={newFriendPhone}
-              onChangeText={setNewFriendPhone}
-              placeholder="(555) 555-5555"
-              placeholderTextColor={colors.textLight}
-              keyboardType="phone-pad"
-            />
-          </View>
-
-          <View style={styles.modalActions}>
-            <Button
-              title="Cancel"
-              onPress={onClose}
-              variant="outline"
-              containerStyle={{ flex: 1, marginHorizontal: 4 }}
-            />
-            <Button
-              title="Add Friend"
-              onPress={handleAdd}
-              containerStyle={{ flex: 1, marginHorizontal: 4 }}
-            />
-          </View>
-        </View>
-      </View>
-    </Modal>
-  );
-};
-
-type GroupItemProps = {
-  group: Group;
-  onPress: () => void;
-};
-
-const GroupItem = ({ group, onPress }: GroupItemProps) => (
-  <TouchableOpacity style={styles.card} onPress={onPress}>
-    <View style={styles.iconContainer}>
-      <Users size={24} color={colors.primary} />
-    </View>
-    <View style={styles.info}>
-      <Text style={styles.name}>{group.name}</Text>
-      <View style={styles.stats}>
-        <View style={styles.stat}>
-          <User size={14} color={colors.textSecondary} />
-          <Text style={styles.statText}>{group.members.length} members</Text>
-        </View>
-        {group.receipts > 0 && (
-          <View style={styles.stat}>
-            <Receipt size={14} color={colors.textSecondary} />
-            <Text style={styles.statText}>{group.receipts} receipts</Text>
-          </View>
-        )}
-      </View>
-    </View>
-    <ChevronRight size={20} color={colors.textLight} />
-  </TouchableOpacity>
-);
-
-type FriendItemProps = {
-  friend: Friend;
-  onPress: () => void;
-};
-
-const FriendItem = ({ friend, onPress }: FriendItemProps) => (
-  <TouchableOpacity style={styles.card} onPress={onPress}>
-    <View style={styles.iconContainer}>
-      <User size={24} color={colors.primary} />
-    </View>
-    <View style={styles.info}>
-      <Text style={styles.name}>{friend.name}</Text>
-      <Text style={styles.phone}>{friend.phone}</Text>
-    </View>
-    <ChevronRight size={20} color={colors.textLight} />
-  </TouchableOpacity>
-);
-
-type SearchBarProps = {
-  value: string;
-  onChangeText: (text: string) => void;
-  placeholder: string;
-};
-
-const SearchBar = ({ value, onChangeText, placeholder }: SearchBarProps) => (
-  <View style={styles.searchContainer}>
-    <Search size={18} color={colors.textSecondary} />
-    <TextInput
-      style={styles.searchInput}
-      placeholder={placeholder}
-      placeholderTextColor={colors.textLight}
-      value={value}
-      onChangeText={onChangeText}
-    />
-    {value.length > 0 && (
-      <TouchableOpacity onPress={() => onChangeText("")}>
-        <X size={18} color={colors.textSecondary} />
-      </TouchableOpacity>
-    )}
-  </View>
-);
 
 export default function GroupsScreen() {
   const router = useRouter();
@@ -479,121 +234,8 @@ const styles = StyleSheet.create({
     color: colors.primary,
     fontWeight: "500",
   },
-  searchContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    backgroundColor: colors.card,
-    borderRadius: 8,
-    paddingHorizontal: 12,
-    marginHorizontal: 16,
-    marginBottom: 16,
-  },
-  searchInput: {
-    flex: 1,
-    paddingVertical: 10,
-    paddingHorizontal: 8,
-    fontSize: 14,
-    color: colors.text,
-  },
   listContent: {
     padding: 16,
     paddingTop: 0,
-  },
-  card: {
-    flexDirection: "row",
-    alignItems: "center",
-    backgroundColor: colors.card,
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 12,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.1,
-    shadowRadius: 2,
-    elevation: 2,
-  },
-  iconContainer: {
-    width: 48,
-    height: 48,
-    borderRadius: 24,
-    backgroundColor: colors.primary + "20",
-    alignItems: "center",
-    justifyContent: "center",
-    marginRight: 16,
-  },
-  info: {
-    flex: 1,
-  },
-  name: {
-    fontSize: 18,
-    fontWeight: "600",
-    color: colors.text,
-    marginBottom: 4,
-  },
-  phone: {
-    fontSize: 14,
-    color: colors.textSecondary,
-  },
-  stats: {
-    flexDirection: "row",
-  },
-  stat: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginRight: 16,
-  },
-  statText: {
-    fontSize: 14,
-    color: colors.textSecondary,
-    marginLeft: 6,
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: "flex-end",
-    backgroundColor: colors.overlay,
-  },
-  modalContent: {
-    backgroundColor: colors.card,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    padding: 20,
-    paddingBottom: Platform.OS === "ios" ? 40 : 20,
-  },
-  modalHeader: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: 20,
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: "600",
-    color: colors.text,
-  },
-  closeButton: {
-    padding: 4,
-  },
-  inputContainer: {
-    marginBottom: 16,
-  },
-  inputLabel: {
-    fontSize: 14,
-    fontWeight: "500",
-    color: colors.textSecondary,
-    marginBottom: 8,
-  },
-  input: {
-    backgroundColor: colors.background,
-    borderRadius: 8,
-    padding: 12,
-    fontSize: 16,
-    color: colors.text,
-    borderWidth: 1,
-    borderColor: colors.border,
-  },
-  modalActions: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginTop: 16,
   },
 });
